hal_adapter: clean up unused utility functions
Remove the direct dependency on the Android Audio
Common Types. Because the IHalAdapterVendorExtension
'av-audio-types-aidl' made dependent on the "android.hardware.audio.core-V2" explicitly.

But unused utility function dependent on the
latest versions which may be different. Hence,
we are removing this unused utility.

Add explicit comment about system property regarding
the interface.



Change-Id: Ie23f13bb335f12b3ef2f6881d152514cd5e4610b